refac: Search Bar Refactor
