+++ instagram/src/components/SearchBar/SearchBar.css
@@ -1,3 +1,4 @@
+/* Searchbar full container */
 .searchbar-environment-container {
 	box-sizing: border-box;
 	background-color: white;
@@ -17,6 +18,7 @@
 	flex-direction: row;
 }
 
+/* Searchbar left part of the container */
 .searchbar-left-container {
 	width: 20%;
 	display: flex;
@@ -34,6 +36,7 @@
 	cursor: pointer;
 }
 
+/* Searchbar middle part of the container */
 .searchbar-middle-container {
 	width: 50%;
 	display: flex;
@@ -53,6 +56,7 @@
 	font-size: 0.9rem;
 }
 
+/* Searchbar right part of the container */
 .searchbar-right-container {
 	width: 20%;
 	display: flex;
